Del get_date.py Update processing.py and widget.py
