+++ src/widget.py
@@ -1,3 +1,5 @@
+from datetime import datetime
+
 from src.masks import get_mask_account, get_mask_card_number
 
 
@@ -19,6 +21,14 @@ def mask_account_card(card_number: str) -> str:
         return f"{card_number[: symbols_count]}{get_mask_account(card_number[- numbers_count:])}"
 
 
+def format_date(date: str) -> str:
+    """Фуккция для преобразования формата даты"""
+    date_get = date[:10]
+    date_format = datetime.strptime(date_get, "%Y-%m-%d")
+    new_date_str = date_format.strftime("%m.%d.%Y")
+    return new_date_str
+
+
 if __name__ == "__main__":
     print(mask_account_card("Maestro 1596837868705199"))
     print(mask_account_card("Счет 64686473678894779589"))
@@ -28,3 +38,4 @@ if __name__ == "__main__":
     print(mask_account_card("Visa Platinum 8990922113665229"))
     print(mask_account_card("Visa Gold 5999414228426353"))
     print(mask_account_card("Счет 73654108430135874305"))
+    print(format_date("2024-03-11T02:26:18.671407"))
